coverage: Move unused-function helpers closer to where they are used
